rename: 페이지 라우팅 및 명칭 정리
