+++ src/App.tsx
@@ -1,17 +1,15 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 
 import MobileView from './layouts/MobileView';
-import Agreement from './pages/Agreement.tsx';
-import Feed from './pages/feed/Feed.page';
-import Feeds from './pages/feed/Feeds.page';
-import KakaoRedirect from './pages/KakaoRedirect.tsx';
-import Login from './pages/Login.tsx';
-import MainLanding from './pages/MainLanding.tsx';
+import Feed from './pages/Feed/Feed.page.tsx';
+import FeedDetail from './pages/Feed/FeedDetail.page.tsx';
+import Agreement from './pages/Login/Agreement.tsx';
+import KakaoRedirect from './pages/Login/KakaoRedirect.tsx';
+import Login from './pages/Login/Login.tsx';
+import SignUp from './pages/Login/SignUp.tsx';
 import More from './pages/profile/More.page.tsx';
 import Profile from './pages/profile/Profile.page.tsx';
 import Setting from './pages/profile/Setting.page.tsx';
-import SignUp from './pages/SignUp.tsx';
-import TestPage from './pages/Test.page';
 import Write from './pages/Write.tsx';
 
 function App() {
@@ -19,12 +17,10 @@ function App() {
     <BrowserRouter>
       <Routes>
         <Route element={<MobileView />}>
-          <Route path="/landing" element={<MainLanding />} />
           <Route path="/login" element={<Login />} />
           <Route path="/oauth/redirected/kakao" element={<KakaoRedirect />} />
-          <Route path="/" element={<TestPage />} />
-          <Route path="/feed" element={<Feeds />} />
-          <Route path="/feed/:id" element={<Feed />} />
+          <Route path="/" element={<Feed />} />
+          <Route path="/feed/:id" element={<FeedDetail />} />
           <Route path="/write" element={<Write />} />
           <Route path="/profile" element={<Profile />} />
           <Route path="/agreement" element={<Agreement />} />
