--- src/pages/MainLanding.tsx
@@ -1,5 +1,0 @@
-const MainLanding = () => {
-  return <div>MainLanding</div>;
-};
-
-export default MainLanding;
